finish alg

+++ AndroidNavigatorLETI/app/src/main/java/com/example/androidnavigatorleti/data/ViewStateParams.kt
@@ -1,8 +1,8 @@
 package com.example.androidnavigatorleti.data
 
 data class ViewStateParams(
-    var minSpeed: Int = 0,
-    var maxSpeed: Int = 0,
+    var minSpeed: Long = 0,
+    var maxSpeed: Long = 0,
     var currentSpeed: Int = 0,
     var currentDistance: Double = 0.0,
     var trafficLightDistance: Double = 0.0
